input types refactor
